red-team/requirements.txt
pyyaml
graphviz
mitmproxy
jsonschema
yamale
